add check_login
