+++ contributions/catalog/controllers/contribute.py
@@ -7,7 +7,7 @@ from flask import (
     Blueprint, render_template, request, session, redirect, url_for
 )
 from requests_oauthlib import OAuth2Session
-
+from .auth import login_required
 from controllers.config import Config as cfg
 from models.contribution_utilities import to_contribution
 from utils import jsonutil
@@ -55,13 +55,7 @@ def result():
         search(query)
         return render_template("contribute/results.html", user=session["name"], result=result)
 
-def login_required(view):
-    @functools.wraps(view)
-    def wrapped_view(**kwargs):
-        if session.get('name') is None:
-            return redirect(url_for('contribute.login'))
-        return view(**kwargs)
-    return wrapped_view
+
 
 
 @bp.route('/create', methods=['GET', "POST"])
